requirements.txt
click
click_help_colors
requests
requests_cache
jinja2
python-dotenv